Move monthly estimated cost methods from resource.py to their respective files
